Add in support for runtime patching on PSOGC.
This also cleans up a few loose ends in the Dreamcast implementation as well.

Please note that PSO Episode I & II Plus do not support patching at all, as it
appears that Sega/Sonic Team removed the code to do so in those versions, due
to the PSOload "exploit". Short of manually patching those games ahead of time
to restore the functionality (which would be silly), there's no way to use any
of this stuff on them.